refactor: worker/job shared 의존성 제거
