+++ skills/spec-review/scripts/spec-review-job.js
@@ -2,7 +2,8 @@
 
 const fs = require('fs');
 const path = require('path');
-const core = require('../../shared/lib/job-core.js');
+const crypto = require('crypto');
+const { spawn } = require('child_process');
 
 const SCRIPT_DIR = __dirname;
 const SKILL_DIR = path.resolve(SCRIPT_DIR, '..');
@@ -10,14 +11,6 @@ const WORKER_PATH = path.join(SCRIPT_DIR, 'spec-review-worker.js');
 
 const SKILL_CONFIG_FILE = path.join(SKILL_DIR, 'spec-review.config.yaml');
 const REPO_CONFIG_FILE = path.join(path.resolve(SKILL_DIR, '../..'), 'spec-review.config.yaml');
-
-const ENTITY_CONFIG = {
-  entityKey: 'reviewer',
-  entitiesDirName: 'reviewers',
-  uiLabel: '[Spec Review]',
-  skillName: 'Spec Review',
-  safeNameFallback: 'reviewer',
-};
 
 const UI_STRINGS = {
   dispatch: {
@@ -30,6 +23,610 @@ const UI_STRINGS = {
     pending: 'Waiting to synthesize',
   },
 };
+
+// ---------------------------------------------------------------------------
+// Pure utility functions
+// ---------------------------------------------------------------------------
+
+function exitWithError(message) {
+  process.stderr.write(`${message}\n`);
+  process.exit(1);
+}
+
+function detectHostRole(skillDir) {
+  const normalized = skillDir.replace(/\\/g, '/');
+  if (normalized.includes('/.claude/skills/')) return 'claude';
+  if (normalized.includes('/.codex/skills/')) return 'codex';
+  return 'unknown';
+}
+
+function normalizeBool(value) {
+  if (value == null) return null;
+  const v = String(value).trim().toLowerCase();
+  if (['1', 'true', 'yes', 'y', 'on'].includes(v)) return true;
+  if (['0', 'false', 'no', 'n', 'off'].includes(v)) return false;
+  return null;
+}
+
+function resolveAutoRole(role, hostRole) {
+  const roleLc = String(role || '').trim().toLowerCase();
+  if (roleLc && roleLc !== 'auto') return roleLc;
+  if (hostRole === 'codex') return 'codex';
+  if (hostRole === 'claude') return 'claude';
+  return 'claude';
+}
+
+function ensureDir(dirPath) {
+  fs.mkdirSync(dirPath, { recursive: true });
+}
+
+function safeFileName(name, fallback) {
+  const cleaned = String(name || '').trim().toLowerCase().replace(/[^a-z0-9_-]+/g, '-');
+  return cleaned || (fallback || 'reviewer');
+}
+
+function atomicWriteJson(filePath, payload) {
+  const tmpPath = `${filePath}.${process.pid}.${crypto.randomBytes(4).toString('hex')}.tmp`;
+  fs.writeFileSync(tmpPath, JSON.stringify(payload, null, 2), 'utf8');
+  fs.renameSync(tmpPath, filePath);
+}
+
+function readJsonIfExists(filePath) {
+  try {
+    if (!fs.existsSync(filePath)) return null;
+    return JSON.parse(fs.readFileSync(filePath, 'utf8'));
+  } catch {
+    return null;
+  }
+}
+
+function sleepMs(ms) {
+  const msNum = Number(ms);
+  if (!Number.isFinite(msNum) || msNum <= 0) return;
+  const sab = new SharedArrayBuffer(4);
+  const view = new Int32Array(sab);
+  Atomics.wait(view, 0, 0, Math.trunc(msNum));
+}
+
+function computeTerminalDoneCount(counts) {
+  const c = counts || {};
+  return (
+    Number(c.done || 0) +
+    Number(c.missing_cli || 0) +
+    Number(c.error || 0) +
+    Number(c.timed_out || 0) +
+    Number(c.canceled || 0)
+  );
+}
+
+function asCodexStepStatus(value) {
+  const v = String(value || '');
+  if (v === 'pending' || v === 'in_progress' || v === 'completed') return v;
+  return 'pending';
+}
+
+// ---------------------------------------------------------------------------
+// Argument parsing
+// ---------------------------------------------------------------------------
+
+function parseArgs(argv) {
+  const args = argv.slice(2);
+  const out = { _: [] };
+  const booleanFlags = new Set([
+    'json',
+    'text',
+    'checklist',
+    'help',
+    'h',
+    'verbose',
+    'include-chairman',
+    'exclude-chairman',
+    'stdin',
+  ]);
+  for (let i = 0; i < args.length; i++) {
+    const a = args[i];
+    if (a === '--') {
+      out._.push(...args.slice(i + 1));
+      break;
+    }
+    if (!a.startsWith('--')) {
+      out._.push(a);
+      continue;
+    }
+
+    const [key, rawValue] = a.split('=', 2);
+    if (rawValue != null) {
+      out[key.slice(2)] = rawValue;
+      continue;
+    }
+
+    const normalizedKey = key.slice(2);
+    if (booleanFlags.has(normalizedKey)) {
+      out[normalizedKey] = true;
+      continue;
+    }
+
+    const next = args[i + 1];
+    if (next == null || next.startsWith('--')) {
+      out[normalizedKey] = true;
+      continue;
+    }
+    out[normalizedKey] = next;
+    i++;
+  }
+  return out;
+}
+
+// ---------------------------------------------------------------------------
+// Wait cursor utilities
+// ---------------------------------------------------------------------------
+
+function parseWaitCursor(value) {
+  const raw = String(value || '').trim();
+  if (!raw) return null;
+  const parts = raw.split(':');
+  const version = parts[0];
+  if (version === 'v1' && parts.length === 4) {
+    const bucketSize = Number(parts[1]);
+    const doneBucket = Number(parts[2]);
+    const isDone = parts[3] === '1';
+    if (!Number.isFinite(bucketSize) || bucketSize <= 0) return null;
+    if (!Number.isFinite(doneBucket) || doneBucket < 0) return null;
+    return { version, bucketSize, dispatchBucket: 0, doneBucket, isDone };
+  }
+  if (version === 'v2' && parts.length === 5) {
+    const bucketSize = Number(parts[1]);
+    const dispatchBucket = Number(parts[2]);
+    const doneBucket = Number(parts[3]);
+    const isDone = parts[4] === '1';
+    if (!Number.isFinite(bucketSize) || bucketSize <= 0) return null;
+    if (!Number.isFinite(dispatchBucket) || dispatchBucket < 0) return null;
+    if (!Number.isFinite(doneBucket) || doneBucket < 0) return null;
+    return { version, bucketSize, dispatchBucket, doneBucket, isDone };
+  }
+  return null;
+}
+
+function formatWaitCursor(bucketSize, dispatchBucket, doneBucket, isDone) {
+  return `v2:${bucketSize}:${dispatchBucket}:${doneBucket}:${isDone ? 1 : 0}`;
+}
+
+function resolveBucketSize(options, total, prevCursor) {
+  const raw = options.bucket != null ? options.bucket : options['bucket-size'];
+
+  if (raw == null || raw === true) {
+    if (prevCursor && prevCursor.bucketSize) return prevCursor.bucketSize;
+  } else {
+    const asString = String(raw).trim().toLowerCase();
+    if (asString !== 'auto') {
+      const num = Number(asString);
+      if (!Number.isFinite(num) || num <= 0) exitWithError(`wait: invalid --bucket: ${raw}`);
+      return Math.trunc(num);
+    }
+  }
+
+  const totalNum = Number(total || 0);
+  if (!Number.isFinite(totalNum) || totalNum <= 0) return 1;
+  return Math.max(1, Math.ceil(totalNum / 5));
+}
+
+// ---------------------------------------------------------------------------
+// Job ID generation
+// ---------------------------------------------------------------------------
+
+function generateJobId() {
+  return `${new Date().toISOString().replace(/[:.]/g, '').replace('T', '-').slice(0, 15)}-${crypto
+    .randomBytes(3)
+    .toString('hex')}`;
+}
+
+// ---------------------------------------------------------------------------
+// Worker spawning
+// ---------------------------------------------------------------------------
+
+function spawnWorkers({ reviewers, workerPath, jobDir, reviewersDir, timeoutSec }) {
+  for (const reviewer of reviewers) {
+    const name = String(reviewer.name);
+    const safeName = safeFileName(name, 'reviewer');
+    const reviewerDir = path.join(reviewersDir, safeName);
+    ensureDir(reviewerDir);
+
+    atomicWriteJson(path.join(reviewerDir, 'status.json'), {
+      reviewer: name,
+      state: 'queued',
+      queuedAt: new Date().toISOString(),
+      command: String(reviewer.command),
+    });
+
+    const workerArgs = [
+      workerPath,
+      '--job-dir',
+      jobDir,
+      '--reviewer',
+      name,
+      '--safe-reviewer',
+      safeName,
+      '--command',
+      String(reviewer.command),
+    ];
+    if (timeoutSec && Number.isFinite(timeoutSec) && timeoutSec > 0) {
+      workerArgs.push('--timeout', String(timeoutSec));
+    }
+
+    const child = spawn(process.execPath, workerArgs, {
+      detached: true,
+      stdio: 'ignore',
+      env: process.env,
+    });
+    child.unref();
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Status computation
+// ---------------------------------------------------------------------------
+
+function computeStatus(jobDir) {
+  const resolvedJobDir = path.resolve(jobDir);
+  if (!fs.existsSync(resolvedJobDir)) exitWithError(`jobDir not found: ${resolvedJobDir}`);
+
+  const jobMeta = readJsonIfExists(path.join(resolvedJobDir, 'job.json'));
+  if (!jobMeta) exitWithError(`job.json not found: ${path.join(resolvedJobDir, 'job.json')}`);
+
+  const reviewersRoot = path.join(resolvedJobDir, 'reviewers');
+  if (!fs.existsSync(reviewersRoot)) exitWithError(`reviewers folder not found: ${reviewersRoot}`);
+
+  const reviewers = [];
+  for (const entry of fs.readdirSync(reviewersRoot)) {
+    const statusPath = path.join(reviewersRoot, entry, 'status.json');
+    const status = readJsonIfExists(statusPath);
+    if (status) reviewers.push({ safeName: entry, ...status });
+  }
+
+  const totals = { queued: 0, running: 0, done: 0, error: 0, missing_cli: 0, timed_out: 0, canceled: 0 };
+  for (const r of reviewers) {
+    const state = String(r.state || 'unknown');
+    if (Object.prototype.hasOwnProperty.call(totals, state)) totals[state]++;
+  }
+
+  const allDone = totals.running === 0 && totals.queued === 0;
+  const overallState = allDone ? 'done' : totals.running > 0 ? 'running' : 'queued';
+
+  return {
+    jobDir: resolvedJobDir,
+    id: jobMeta.id || null,
+    chairmanRole: jobMeta.chairmanRole || null,
+    overallState,
+    counts: { total: reviewers.length, ...totals },
+    specName: jobMeta.specName || null,
+    reviewers: reviewers
+      .map((r) => ({
+        reviewer: r.reviewer,
+        state: r.state,
+        startedAt: r.startedAt || null,
+        finishedAt: r.finishedAt || null,
+        exitCode: r.exitCode != null ? r.exitCode : null,
+        message: r.message || null,
+      }))
+      .sort((a, b) => String(a.reviewer).localeCompare(String(b.reviewer))),
+  };
+}
+
+// ---------------------------------------------------------------------------
+// UI payload
+// ---------------------------------------------------------------------------
+
+function buildUiPayload(statusPayload) {
+  const counts = statusPayload.counts || {};
+  const done = computeTerminalDoneCount(counts);
+  const total = Number(counts.total || 0);
+  const isDone = String(statusPayload.overallState || '') === 'done';
+
+  const queued = Number(counts.queued || 0);
+  const running = Number(counts.running || 0);
+
+  const reviewersArray = Array.isArray(statusPayload.reviewers) ? statusPayload.reviewers : [];
+  const sortedReviewers = reviewersArray
+    .map((r) => ({
+      entity: r && r.reviewer != null ? String(r.reviewer) : '',
+      state: r && r.state != null ? String(r.state) : 'unknown',
+      exitCode: r && r.exitCode != null ? r.exitCode : null,
+    }))
+    .filter((r) => r.entity)
+    .sort((a, b) => a.entity.localeCompare(b.entity));
+
+  const terminalStates = new Set(['done', 'missing_cli', 'error', 'timed_out', 'canceled']);
+  const dispatchStatus = asCodexStepStatus(isDone ? 'completed' : queued > 0 ? 'in_progress' : 'completed');
+  let hasInProgress = dispatchStatus === 'in_progress';
+
+  const reviewerSteps = sortedReviewers.map((r) => {
+    const state = r.state || 'unknown';
+    const isTerminal = terminalStates.has(state);
+
+    let status;
+    if (isTerminal) {
+      status = 'completed';
+    } else if (!hasInProgress && running > 0 && state === 'running') {
+      status = 'in_progress';
+      hasInProgress = true;
+    } else {
+      status = 'pending';
+    }
+
+    const label = `[Spec Review] Ask ${r.entity}`;
+    return { label, status: asCodexStepStatus(status) };
+  });
+
+  const synthStatus = asCodexStepStatus(isDone ? (hasInProgress ? 'pending' : 'in_progress') : 'pending');
+
+  const codexPlan = [
+    { step: '[Spec Review] Prompt dispatch', status: dispatchStatus },
+    ...reviewerSteps.map((s) => ({ step: s.label, status: s.status })),
+    { step: '[Spec Review] Synthesize', status: synthStatus },
+  ];
+
+  const claudeTodos = [
+    {
+      content: '[Spec Review] Prompt dispatch',
+      status: dispatchStatus,
+      activeForm: dispatchStatus === 'completed'
+        ? UI_STRINGS.dispatch.completed
+        : UI_STRINGS.dispatch.inProgress,
+    },
+    ...reviewerSteps.map((s) => ({
+      content: s.label,
+      status: s.status,
+      activeForm: s.status === 'completed' ? 'Finished' : 'Awaiting response',
+    })),
+    {
+      content: '[Spec Review] Synthesize',
+      status: synthStatus,
+      activeForm:
+        synthStatus === 'completed'
+          ? UI_STRINGS.synthesize.completed
+          : synthStatus === 'in_progress'
+            ? UI_STRINGS.synthesize.inProgress
+            : UI_STRINGS.synthesize.pending,
+    },
+  ];
+
+  return {
+    progress: { done, total, overallState: String(statusPayload.overallState || '') },
+    codex: { update_plan: { plan: codexPlan } },
+    claude: { todo_write: { todos: claudeTodos } },
+  };
+}
+
+// ---------------------------------------------------------------------------
+// Wait payload
+// ---------------------------------------------------------------------------
+
+function asWaitPayload(statusPayload) {
+  const reviewersArray = Array.isArray(statusPayload.reviewers) ? statusPayload.reviewers : [];
+
+  return {
+    jobDir: statusPayload.jobDir,
+    id: statusPayload.id,
+    chairmanRole: statusPayload.chairmanRole,
+    overallState: statusPayload.overallState,
+    counts: statusPayload.counts,
+    specName: statusPayload.specName,
+    reviewers: reviewersArray.map((r) => ({
+      reviewer: r.reviewer,
+      state: r.state,
+      exitCode: r.exitCode != null ? r.exitCode : null,
+      message: r.message || null,
+    })),
+    ui: buildUiPayload(statusPayload),
+  };
+}
+
+// ---------------------------------------------------------------------------
+// Command implementations
+// ---------------------------------------------------------------------------
+
+function cmdStatus(options, jobDir) {
+  const payload = computeStatus(jobDir);
+
+  const wantChecklist = Boolean(options.checklist) && !options.json;
+  if (wantChecklist) {
+    const done = computeTerminalDoneCount(payload.counts);
+    const headerId = payload.id ? ` (${payload.id})` : '';
+    const extraInfo = payload.specName ? ` [spec: ${payload.specName}]` : '';
+    process.stdout.write(`Spec Review${headerId}${extraInfo}\n`);
+    process.stdout.write(
+      `Progress: ${done}/${payload.counts.total} done  (running ${payload.counts.running}, queued ${payload.counts.queued})\n`
+    );
+    for (const r of payload.reviewers) {
+      const state = String(r.state || '');
+      const mark =
+        state === 'done'
+          ? '[x]'
+          : state === 'running' || state === 'queued'
+            ? '[ ]'
+            : state
+              ? '[!]'
+              : '[ ]';
+      const exitInfo = r.exitCode != null ? ` (exit ${r.exitCode})` : '';
+      process.stdout.write(`${mark} ${r.reviewer} \u2014 ${state}${exitInfo}\n`);
+    }
+    return;
+  }
+
+  const wantText = Boolean(options.text) && !options.json;
+  if (wantText) {
+    const done = computeTerminalDoneCount(payload.counts);
+    process.stdout.write(`reviewers ${done}/${payload.counts.total} done; running=${payload.counts.running} queued=${payload.counts.queued}\n`);
+    if (options.verbose) {
+      for (const r of payload.reviewers) {
+        process.stdout.write(`- ${r.reviewer}: ${r.state}${r.exitCode != null ? ` (exit ${r.exitCode})` : ''}\n`);
+      }
+    }
+    return;
+  }
+
+  process.stdout.write(`${JSON.stringify(payload, null, 2)}\n`);
+}
+
+function cmdWait(options, jobDir) {
+  const resolvedJobDir = path.resolve(jobDir);
+  const cursorFilePath = path.join(resolvedJobDir, '.wait_cursor');
+  const prevCursorRaw =
+    options.cursor != null
+      ? String(options.cursor)
+      : fs.existsSync(cursorFilePath)
+        ? String(fs.readFileSync(cursorFilePath, 'utf8')).trim()
+        : '';
+  const prevCursor = parseWaitCursor(prevCursorRaw);
+
+  const intervalMsRaw = options['interval-ms'] != null ? options['interval-ms'] : 250;
+  const intervalMs = Math.max(50, Math.trunc(Number(intervalMsRaw)));
+  if (!Number.isFinite(intervalMs) || intervalMs <= 0) exitWithError(`wait: invalid --interval-ms: ${intervalMsRaw}`);
+
+  const timeoutMsRaw = options['timeout-ms'] != null ? options['timeout-ms'] : 0;
+  const timeoutMs = Math.trunc(Number(timeoutMsRaw));
+  if (!Number.isFinite(timeoutMs) || timeoutMs < 0) exitWithError(`wait: invalid --timeout-ms: ${timeoutMsRaw}`);
+
+  let payload = computeStatus(jobDir);
+  const bucketSize = resolveBucketSize(options, payload.counts.total, prevCursor);
+
+  const doneCount = computeTerminalDoneCount(payload.counts);
+  const isDone = payload.overallState === 'done';
+  const total = Number(payload.counts.total || 0);
+  const queued = Number(payload.counts.queued || 0);
+  const dispatchBucket = queued === 0 && total > 0 ? 1 : 0;
+  const doneBucket = Math.floor(doneCount / bucketSize);
+  const cursor = formatWaitCursor(bucketSize, dispatchBucket, doneBucket, isDone);
+
+  if (!prevCursor) {
+    fs.writeFileSync(cursorFilePath, cursor, 'utf8');
+    process.stdout.write(`${JSON.stringify({ ...asWaitPayload(payload), cursor }, null, 2)}\n`);
+    return;
+  }
+
+  const start = Date.now();
+  while (cursor === prevCursorRaw) {
+    if (timeoutMs > 0 && Date.now() - start >= timeoutMs) break;
+    sleepMs(intervalMs);
+    payload = computeStatus(jobDir);
+    const d = computeTerminalDoneCount(payload.counts);
+    const doneFlag = payload.overallState === 'done';
+    const totalCount = Number(payload.counts.total || 0);
+    const queuedCount = Number(payload.counts.queued || 0);
+    const dispatchB = queuedCount === 0 && totalCount > 0 ? 1 : 0;
+    const doneB = Math.floor(d / bucketSize);
+    const nextCursor = formatWaitCursor(bucketSize, dispatchB, doneB, doneFlag);
+    if (nextCursor !== prevCursorRaw) {
+      fs.writeFileSync(cursorFilePath, nextCursor, 'utf8');
+      process.stdout.write(`${JSON.stringify({ ...asWaitPayload(payload), cursor: nextCursor }, null, 2)}\n`);
+      return;
+    }
+  }
+
+  const finalPayload = computeStatus(jobDir);
+  const finalDone = computeTerminalDoneCount(finalPayload.counts);
+  const finalDoneFlag = finalPayload.overallState === 'done';
+  const finalTotal = Number(finalPayload.counts.total || 0);
+  const finalQueued = Number(finalPayload.counts.queued || 0);
+  const finalDispatchBucket = finalQueued === 0 && finalTotal > 0 ? 1 : 0;
+  const finalDoneBucket = Math.floor(finalDone / bucketSize);
+  const finalCursor = formatWaitCursor(bucketSize, finalDispatchBucket, finalDoneBucket, finalDoneFlag);
+  fs.writeFileSync(cursorFilePath, finalCursor, 'utf8');
+  process.stdout.write(`${JSON.stringify({ ...asWaitPayload(finalPayload), cursor: finalCursor }, null, 2)}\n`);
+}
+
+function cmdResults(options, jobDir) {
+  const resolvedJobDir = path.resolve(jobDir);
+  const jobMeta = readJsonIfExists(path.join(resolvedJobDir, 'job.json'));
+  const reviewersRoot = path.join(resolvedJobDir, 'reviewers');
+
+  const reviewers = [];
+  if (fs.existsSync(reviewersRoot)) {
+    for (const entry of fs.readdirSync(reviewersRoot)) {
+      const statusPath = path.join(reviewersRoot, entry, 'status.json');
+      const outputPath = path.join(reviewersRoot, entry, 'output.txt');
+      const errorPath = path.join(reviewersRoot, entry, 'error.txt');
+      const status = readJsonIfExists(statusPath);
+      if (!status) continue;
+      const output = fs.existsSync(outputPath) ? fs.readFileSync(outputPath, 'utf8') : '';
+      const stderr = fs.existsSync(errorPath) ? fs.readFileSync(errorPath, 'utf8') : '';
+      reviewers.push({ safeName: entry, ...status, output, stderr });
+    }
+  }
+
+  if (options.json) {
+    process.stdout.write(
+      `${JSON.stringify(
+        {
+          jobDir: resolvedJobDir,
+          id: jobMeta ? jobMeta.id : null,
+          specName: jobMeta ? jobMeta.specName : null,
+          prompt: fs.existsSync(path.join(resolvedJobDir, 'prompt.txt'))
+            ? fs.readFileSync(path.join(resolvedJobDir, 'prompt.txt'), 'utf8')
+            : null,
+          reviewers: reviewers
+            .map((r) => ({
+              reviewer: r.reviewer,
+              state: r.state,
+              exitCode: r.exitCode != null ? r.exitCode : null,
+              message: r.message || null,
+              output: r.output,
+              stderr: r.stderr,
+            }))
+            .sort((a, b) => String(a.reviewer).localeCompare(String(b.reviewer))),
+        },
+        null,
+        2
+      )}\n`
+    );
+    return;
+  }
+
+  for (const r of reviewers.sort((a, b) => String(a.reviewer).localeCompare(String(b.reviewer)))) {
+    process.stdout.write(`\n=== ${r.reviewer} (${r.state}) ===\n`);
+    if (r.message) process.stdout.write(`${r.message}\n`);
+    process.stdout.write(r.output || '');
+    if (!r.output && r.stderr) {
+      process.stdout.write('\n');
+      process.stdout.write(r.stderr);
+    }
+    process.stdout.write('\n');
+  }
+}
+
+function cmdStop(_options, jobDir) {
+  const resolvedJobDir = path.resolve(jobDir);
+  const reviewersRoot = path.join(resolvedJobDir, 'reviewers');
+  if (!fs.existsSync(reviewersRoot)) exitWithError(`No reviewers folder found: ${reviewersRoot}`);
+
+  let stoppedAny = false;
+  for (const entry of fs.readdirSync(reviewersRoot)) {
+    const statusPath = path.join(reviewersRoot, entry, 'status.json');
+    const status = readJsonIfExists(statusPath);
+    if (!status) continue;
+    if (status.state !== 'running') continue;
+    if (!status.pid) continue;
+
+    try {
+      process.kill(Number(status.pid), 'SIGTERM');
+      stoppedAny = true;
+    } catch {
+      // ignore
+    }
+  }
+
+  process.stdout.write(stoppedAny ? `stop: sent SIGTERM to running reviewers\n` : `stop: no running reviewers\n`);
+}
+
+function cmdClean(_options, jobDir) {
+  const resolvedJobDir = path.resolve(jobDir);
+  fs.rmSync(resolvedJobDir, { recursive: true, force: true });
+  process.stdout.write(`cleaned: ${resolvedJobDir}\n`);
+}
+
+// ---------------------------------------------------------------------------
+// Spec-review-specific config parsing
+// ---------------------------------------------------------------------------
 
 function resolveDefaultConfigFile() {
   if (fs.existsSync(SKILL_CONFIG_FILE)) return SKILL_CONFIG_FILE;
@@ -65,17 +662,17 @@ function parseSpecReviewConfig(configPath) {
     parsed = YAML.parse(fs.readFileSync(configPath, 'utf8'));
   } catch (error) {
     const message = error && error.message ? error.message : String(error);
-    core.exitWithError(`Invalid YAML in ${configPath}: ${message}`);
+    exitWithError(`Invalid YAML in ${configPath}: ${message}`);
   }
 
   if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) {
-    core.exitWithError(`Invalid config in ${configPath}: expected a YAML mapping/object at the document root`);
+    exitWithError(`Invalid config in ${configPath}: expected a YAML mapping/object at the document root`);
   }
   if (!parsed['spec-review']) {
-    core.exitWithError(`Invalid config in ${configPath}: missing required top-level key 'spec-review:'`);
+    exitWithError(`Invalid config in ${configPath}: missing required top-level key 'spec-review:'`);
   }
   if (typeof parsed['spec-review'] !== 'object' || Array.isArray(parsed['spec-review'])) {
-    core.exitWithError(`Invalid config in ${configPath}: 'spec-review' must be a mapping/object`);
+    exitWithError(`Invalid config in ${configPath}: 'spec-review' must be a mapping/object`);
   }
 
   const merged = {
@@ -91,28 +688,28 @@ function parseSpecReviewConfig(configPath) {
 
   if (specReview.chairman != null) {
     if (typeof specReview.chairman !== 'object' || Array.isArray(specReview.chairman)) {
-      core.exitWithError(`Invalid config in ${configPath}: 'spec-review.chairman' must be a mapping/object`);
+      exitWithError(`Invalid config in ${configPath}: 'spec-review.chairman' must be a mapping/object`);
     }
     merged['spec-review'].chairman = { ...merged['spec-review'].chairman, ...specReview.chairman };
   }
 
   if (Object.prototype.hasOwnProperty.call(specReview, 'reviewers')) {
     if (!Array.isArray(specReview.reviewers)) {
-      core.exitWithError(`Invalid config in ${configPath}: 'spec-review.reviewers' must be a list/array`);
+      exitWithError(`Invalid config in ${configPath}: 'spec-review.reviewers' must be a list/array`);
     }
     merged['spec-review'].reviewers = specReview.reviewers;
   }
 
   if (specReview.context != null) {
     if (typeof specReview.context !== 'object' || Array.isArray(specReview.context)) {
-      core.exitWithError(`Invalid config in ${configPath}: 'spec-review.context' must be a mapping/object`);
+      exitWithError(`Invalid config in ${configPath}: 'spec-review.context' must be a mapping/object`);
     }
     merged['spec-review'].context = { ...merged['spec-review'].context, ...specReview.context };
   }
 
   if (specReview.settings != null) {
     if (typeof specReview.settings !== 'object' || Array.isArray(specReview.settings)) {
-      core.exitWithError(`Invalid config in ${configPath}: 'spec-review.settings' must be a mapping/object`);
+      exitWithError(`Invalid config in ${configPath}: 'spec-review.settings' must be a mapping/object`);
     }
     merged['spec-review'].settings = { ...merged['spec-review'].settings, ...specReview.settings };
   }
@@ -185,7 +782,6 @@ function parseYamlSimple(configPath, fallback) {
 // ---------------------------------------------------------------------------
 
 function findProjectRoot() {
-  // Look for .omt directory starting from skill directory and going up
   let current = SKILL_DIR;
   const root = path.parse(current).root;
 
@@ -195,7 +791,6 @@ function findProjectRoot() {
       return current;
     }
 
-    // Also check for .git as fallback project root indicator
     const gitDir = path.join(current, '.git');
     if (fs.existsSync(gitDir)) {
       return current;
@@ -204,7 +799,6 @@ function findProjectRoot() {
     current = path.dirname(current);
   }
 
-  // Fallback: try common patterns
   const normalized = SKILL_DIR.replace(/\\/g, '/');
   const skillsMatch = normalized.match(/^(.+?)\/.claude\/skills\//);
   if (skillsMatch) {
@@ -214,7 +808,6 @@ function findProjectRoot() {
   return null;
 }
 
-// Gather spec context from .omt/specs/ directory
 function gatherSpecContext(specName, config) {
   const parts = [];
   const projectRoot = findProjectRoot();
@@ -232,7 +825,6 @@ function gatherSpecContext(specName, config) {
 
   const files = [];
 
-  // 1. Gather ALL *.md files from shared_context_dir
   if (fs.existsSync(contextDir)) {
     try {
       const contextFiles = fs.readdirSync(contextDir)
@@ -254,7 +846,6 @@ function gatherSpecContext(specName, config) {
     }
   }
 
-  // 2. Gather spec.md from {specs_dir}/{specName}/
   const specPath = path.join(specDir, 'spec.md');
   if (fs.existsSync(specPath)) {
     try {
@@ -266,7 +857,6 @@ function gatherSpecContext(specName, config) {
     }
   }
 
-  // 3. Gather decision records from {specs_dir}/{specName}/records/
   const recordsDir = path.join(specDir, 'records');
   if (fs.existsSync(recordsDir)) {
     try {
@@ -296,24 +886,8 @@ function gatherSpecContext(specName, config) {
 }
 
 // ---------------------------------------------------------------------------
-// Spec-review-specific command implementations
+// Spec-review-specific start command
 // ---------------------------------------------------------------------------
-
-function computeStatus(jobDir) {
-  return core.computeStatusPayload(jobDir, ENTITY_CONFIG, (jobMeta) => ({
-    specName: jobMeta.specName || null,
-  }));
-}
-
-function buildSpecReviewUiPayload(statusPayload) {
-  return core.buildUiPayload(statusPayload, ENTITY_CONFIG, UI_STRINGS);
-}
-
-function specReviewAsWaitPayload(statusPayload) {
-  return core.asWaitPayload(statusPayload, ENTITY_CONFIG, buildSpecReviewUiPayload, (sp) => ({
-    specName: sp.specName,
-  }));
-}
 
 function printHelp() {
   process.stdout.write(`Spec Review (job mode)
@@ -340,18 +914,18 @@ function cmdStart(options, prompt) {
   const jobsDir =
     options['jobs-dir'] || process.env.SPEC_REVIEW_JOBS_DIR || path.join(SKILL_DIR, '.jobs');
 
-  core.ensureDir(jobsDir);
+  ensureDir(jobsDir);
 
-  const hostRole = core.detectHostRole(SKILL_DIR);
+  const hostRole = detectHostRole(SKILL_DIR);
   const config = parseSpecReviewConfig(configPath);
   const chairmanRoleRaw = options.chairman || process.env.SPEC_REVIEW_CHAIRMAN || config['spec-review'].chairman.role || 'auto';
-  const chairmanRole = core.resolveAutoRole(chairmanRoleRaw, hostRole);
+  const chairmanRole = resolveAutoRole(chairmanRoleRaw, hostRole);
 
   const includeChairman = Boolean(options['include-chairman']);
   const excludeChairmanOverride =
     options['exclude-chairman'] != null ? true : options['include-chairman'] != null ? false : null;
 
-  const excludeSetting = core.normalizeBool(config['spec-review'].settings.exclude_chairman_from_reviewers);
+  const excludeSetting = normalizeBool(config['spec-review'].settings.exclude_chairman_from_reviewers);
   const excludeChairmanFromReviewers =
     excludeChairmanOverride != null ? excludeChairmanOverride : excludeSetting != null ? excludeSetting : true;
 
@@ -367,14 +941,12 @@ function cmdStart(options, prompt) {
     return true;
   });
 
-  // Handle spec context
   const specName = options.spec || null;
   let specContext = { context: '', files: [] };
   if (specName) {
     specContext = gatherSpecContext(specName, config);
   }
 
-  // Build final prompt with spec context
   let finalPrompt = prompt;
   if (specContext.context) {
     finalPrompt = `# Spec Review Context
@@ -388,10 +960,10 @@ ${specContext.context}
 ${prompt}`;
   }
 
-  const jobId = core.generateJobId();
+  const jobId = generateJobId();
   const jobDir = path.join(jobsDir, `spec-review-${jobId}`);
   const reviewersDir = path.join(jobDir, 'reviewers');
-  core.ensureDir(reviewersDir);
+  ensureDir(reviewersDir);
 
   fs.writeFileSync(path.join(jobDir, 'prompt.txt'), String(finalPrompt), 'utf8');
 
@@ -414,14 +986,13 @@ ${prompt}`;
       color: r.color ? String(r.color) : null,
     })),
   };
-  core.atomicWriteJson(path.join(jobDir, 'job.json'), jobMeta);
+  atomicWriteJson(path.join(jobDir, 'job.json'), jobMeta);
 
-  core.spawnWorkers({
-    entities: reviewers,
-    entityConfig: ENTITY_CONFIG,
+  spawnWorkers({
+    reviewers,
     workerPath: WORKER_PATH,
     jobDir,
-    entitiesDir: reviewersDir,
+    reviewersDir,
     timeoutSec,
   });
 
@@ -436,22 +1007,58 @@ ${prompt}`;
 // Main
 // ---------------------------------------------------------------------------
 
-const STATUS_LABELS = {
-  checklistExtraFn: (payload) => payload.specName ? ` [spec: ${payload.specName}]` : '',
-};
-
 function main() {
-  core.mainRouter({
-    printHelp,
-    cmdStart,
-    cmdStatus: (options, jobDir) => core.cmdStatus(options, jobDir, ENTITY_CONFIG, computeStatus, STATUS_LABELS),
-    cmdWait: (options, jobDir) => core.cmdWait(options, jobDir, ENTITY_CONFIG, computeStatus, specReviewAsWaitPayload),
-    cmdResults: (options, jobDir) => core.cmdResults(options, jobDir, ENTITY_CONFIG, (jobMeta) => ({
-      specName: jobMeta ? jobMeta.specName : null,
-    })),
-    cmdStop: (options, jobDir) => core.cmdStop(options, jobDir, ENTITY_CONFIG),
-    cmdClean: (options, jobDir) => core.cmdClean(options, jobDir),
-  });
+  const options = parseArgs(process.argv);
+  const [command, ...rest] = options._;
+
+  if (!command || options.help || options.h) {
+    printHelp();
+    return;
+  }
+
+  if (command === 'start') {
+    let prompt;
+    if (options.stdin) {
+      prompt = fs.readFileSync(0, 'utf8');
+    } else {
+      prompt = rest.join(' ').trim();
+    }
+    if (!prompt) exitWithError('start: missing prompt');
+    cmdStart(options, prompt);
+    return;
+  }
+  if (command === 'status') {
+    const jobDir = rest[0];
+    if (!jobDir) exitWithError('status: missing jobDir');
+    cmdStatus(options, jobDir);
+    return;
+  }
+  if (command === 'wait') {
+    const jobDir = rest[0];
+    if (!jobDir) exitWithError('wait: missing jobDir');
+    cmdWait(options, jobDir);
+    return;
+  }
+  if (command === 'results') {
+    const jobDir = rest[0];
+    if (!jobDir) exitWithError('results: missing jobDir');
+    cmdResults(options, jobDir);
+    return;
+  }
+  if (command === 'stop') {
+    const jobDir = rest[0];
+    if (!jobDir) exitWithError('stop: missing jobDir');
+    cmdStop(options, jobDir);
+    return;
+  }
+  if (command === 'clean') {
+    const jobDir = rest[0];
+    if (!jobDir) exitWithError('clean: missing jobDir');
+    cmdClean(options, jobDir);
+    return;
+  }
+
+  exitWithError(`Unknown command: ${command}`);
 }
 
 if (require.main === module) {
